refactor: Test UserFixtures 폴더명 변경
