+++ module-api/src/test/groovy/com/example/moduleapi/fixture/user/UserFixtures.groovy
@@ -1,4 +1,4 @@
-package com.example.moduleapi.fixture.UserFixtures
+package com.example.moduleapi.fixture.user
 
 import com.example.moduledomain.domain.user.User
